This resolved #4213
Color names are used for lines in 2D curve plots. The colors for Critical Edge Selection and for all curves in the reflectivity (I vs Q and I vs lambda perpendicular) are specified in the configuration file. The color table for data plots is also specified in the configuration file.
The color for lines drawn on data plots are still specified in terms of the color table as a numeric value between 0 and 255. Examples of values for these will be specified in the documentation for the color tables 5 (standard gamma II) and 13 (rainbow). 

To remove problem of window being erased, the parameter RETAIN=2 was added to the WIDGET_DRAW call in a number of routines. This forces IDL to deal with redrawing the screen if something covers the window. We were getting some very variable behavior, but that appears to all be fixed now.

Richard Ward
March 8, 2010