This resolved #4212
Corrections were made to the saving to a file (in ref_off_spec_reduce_sangle) and loading from this file (in ref_off_spec_eventcb) of the RefPix values.
This was tested and corrections were made to branch around reading of the file if it did not exist.

A label was added to the MakeGuiStep2 to indicate to the user that, if no shifting of the data was required, the user could continue on directly to the "Scaling" step.

Richard Ward
March 8, 2010