This resolved #4252
Fixed problem with display of 2nd file in the SCALING plot. On clicking "Scaling Other Files" all but the reference data are scaled and plotted using the initial value of the scaling factor in the text box. That value was 0.0, so that the scaled values of the 2nd set of data were being set to INF. This did not affect the 3rd set of data for some reason. 
To correct this, the initial value of the scaling (set to 1.0) in MakeGuiStep6.pro was properly set. This corrected the problem.

I also corrected an in advertent error introduced in the routine step6_realign_data in the code section ref_off_spec_step6.pro. A couple letters had gotten chopped off an array name causing an error. But the location of the error was not apparent, since the CATCH function was used to catch errors and print those to the application. This particular error was not caught and not displayed anywhere.