This resolved #4508
Modifications to the RECAP step (renamed Scaling-2D) to use the same plotting capability as in Step 4, with the same ability to change the [xmin,xmax] and [ymin,ymax] of the plot.
Further clean up of Step5 is required for completion of this modification.

In addition, changes were made to the Scaling Step (now called Scaling-1D). The entire range of the selected curve for the CE selection (principal dataset) was not being captured. This has been fixed, although I am not getting a error which could be related to changes in this step when running a recent magnetism reflectometry dataset. This also needs further investigation.

Other small changes were made.

Richard Ward
May 7, 2010