This resolved #4831
Use a single directory path (ascii_path) to store both the ROI files (*.dat), the RefPix file (*RefPix.txt), and the reduced data files (*.txt). The user can enter a different directory path for the output files (working_path) at Step 5. This new path is passed automatically to Step 6. If the user does not select a new output path, this defaults to the same location as the reduced data (ascii_path). 

The user can alter the intial location for ascii_path in the first screen of the application, where the user selects the instrument type (REF_M, REF_L) and screen resolution. The default
location for ascii_path is ~/result/, which is now imbedded in the code for the first screen, and passed to the main routine where the global array is set.

RC Ward
26 July 2010