This resolved #4833
This fixed the computation of the other spin states in Step 6 (Output Step), which was not being done properly because pixel_offset was not being used properly. This was noted on plots of the 2D scaled output for spin state On_Off. The correction has been made. A small number of other alterations were made in the wording in the titles used for DIALOG_PICKFILE routines in Steps 5 and 6. 
RC Ward  1 Aug 2010