This resolved #4835
These fixes were done to prepare the code for 1.5.1 release:
Old rmd files are now also deleted when the txt (Ascii) files are deleted prior to launching the refred_lp remote reduction code. This will remove any imbiguity. 
Changed the position of the linear/log buttons on the Reduction step screen.
Change the images used for selecting the spin states for doing the reduction processing and for alerting the the JobMananger is on the way.
Set ymax limit to 10. in the scaling step after maing the CE selection. This saves a step for user having to do so in the ymax box.
Noted that TOF cuttoffs can be done in REF_L by setting these in the CONFIG file. Since these are specific to the instrument, they are likely to be set only once and not to be set interactively. 
Do not have interactice capability now for REF_L.
Introduce a new splicing alternative in the Step 5. The original apporach used the MAX of the values in the overlap region. The alternative (which is now the default) overrides the higher Q dataset over the previous one. This is actually done as a toggle on the initial Step 5 screen. Allows user to toggle between these to splicing approaches. 
RC Ward  13 Aug 2010