This resolves #5278
The following changes were made: 
1. Code creates only one RefPix file instead of one file for each spins state. The form is REF_M_Run#_RefPix.txt
2. RefPix values can now be changed in the Shifting step and saved to the RefPix.txt output file
If the CANCEL button is clicked, the original values of RefPix are returned to the screen, and written to the RefPix
output file. 
3. To acomplish 2.) the number of digits shown for Refpix value in the Shifting step was expanded to show 1/2 pixel values.
This was changes in two locations on the screen: where the y pixel values are displayed and in the input text box for entering
modification to RefPix values for each dataset.
4. On the FRONT screen, the code was fixed so the user can type in the results directory without the following "/". 
The trailing "/" is added the program if not present. 
RC Ward
16 Jan 2011